Refactor command management system
This commit includes lots of changes to refactor the way commands are handled by the system. It removes the poorly flexible and extensible
parser in favor of a new system that respects the Open-Closed SOLID principle. With this new system you can easily add new commands
without having to modify the existing logic.