Create new InvalidArgumentExceptionClass as base class of all defined custom exceptions and catch it in Application::run() method
This commit creates a new exception that will be the base class of all the custom exceptions already defined in the application.
Then we catch this new exception in the Application::run() method instead of catching the fully generic and meaningless php
Exception class. In this way, we are catching only the application related exceptions and letting the others bubble up.