Move instanciation of application collaborators into the Application class constructor
Previously we had to instanciate the application collaborators a pass them to the Application class at the
moment of its instanciation. Now all that logic has been moved into the application class constructor to
encapsulate it and to avoid boilerplate. However, I've added setter methods for those collaborators to be
able to change the application behaviour at runtime. This could be quite useful for testing purposes.